Improve iptables rules
Add the option to select which interface to filter.

Improve iptables rules to avoid receiving packet twice. This was
happening when locally a user try to connect to a local service. The
packet was captured by the OUTPUT target and also by the PREROUTING. We
added a mark on the packet to avoid this problem.