Fix ack number offset
Due to tcp modificatons, the ack number need to be offset. Before we had
the assumption that only full packet could be acked. Due to the change
where tcp is now able to split a big packet, we can receive ack that are
not only at packet boundary. We then need to iterate the modifications done
to be able to find the correct offset to apply.