Add RUNPATH to build applications
The RUNPATH allow the LD_LIBRARY_PATH paths to be search before the
RPATH one. This is needed to make sure that when setting different paths
in LD_LIBRARY_PATH, thos library will be searched before. It happens
when haka is installed on the current computer but another one (from a
localinstall for instance) is started. The localinstall binaries need to
use the localinstall libraries.