Add Lua VM interruption
In order to implement timer interruption, we need a way to execute on a
signal some Lua code. The implementation allows to interrupt a VM that
will then call the Lua code. This interruption is possible from a
signal. However, it is not thread safe and need to be called from the
same thread.