Fix threading issue with lua state
When running in multi-threads, the Lua state used to be initialized by
the main thread. This leads to some major problem with timers that are
going to signal the thread that have created them. To avoid this, the
Lua state initialization need to be moved inside the thread. However,
to keep the errors and information displayed during the initialization,
the daemonize method have been updated.