Fix state machine issues
The state machine could have some stack overflow issue if trying to
access an element that does not exists.

Also the test has been change to avoid rely on the debug output as it is
now filtered out by the output formatter to avoid having to much effect
on all tests.