Improve tcp state handing
Reimplement the tcp state handling using the state machine. It is also a
lot more precise compared to what we used to have. It will still need
some polish to check all possible transitions as well as the timeouts.

Updated the test to take into account the delay connection closing and
dropping.