Improve tests
The packet creation tests could end up in an infinite loop in case of
failure. To avoid this, we now limit the number of packet to be sent. If
the test send to many packets it will still be in failure state, however
it will not enter the infinite loop waiting for the test timeout.