Fix LuaJit thread cancel issue
The LuaJit compilation was not correct, the custom options were not
taken into account. The makefile has been fixed to use our compilation
options.

Also we had a major issue with LuaJit, when cancelling the threads (in
response to a ^C for instance), if at this same time the Lua VM was
inside a protected call (pcall, xpcall...), then the LuaJit code was
catching the stack unwinding thus aborting the thread cancelation. This
problem was only happening on x86 as the unwinding method is not the
same depending on the architecture. To fix this, the option
LUAJIT_UNWIND_EXTERNAL needs to be added along with the gcc option
-funwind-tables. This mode is the mode used in x64 and works better in
our case.