Fix table order issue in tests
The table element order from a call to pairs() does not give element in
the same order on Lua 5.2. Each run of the program will have a different
order due to a seed used to initialize the hash function. To be able to
check for the output in our tests, we need to sort the table element.