Fix interactive mode prompt issue
With a more recent libedit, the colors in the prompt are not correctly
printed resulting in strange ']0m' characters. To fix this, we will have
to use the real libedit API not the readline mapping built on top of it
and escape the color codes by a special character. The examples in the
libedit code shows how to do it. The problem is visible on Fedora 20.