Fix lua_error calls
lua_error will do a longjump and the value at the top of the stack will
be the error message. We have a problem where an invalid error message
was reported in some case. This is probably a result of an invalid use
of lua_error. All call to this function have been checked and reinforced
to avoid possible issues. The problem is not visible anymore after this
fix.