Fix iterator issue in coroutine mode
With a blocking iterator it is not always possible to copy it before
calling sub('available'). This copy might end up being at the end of the
vbuffer thus being after the sub returned by sub('available'). To fix
this we added a new function wait() to allow to block in order to get
some more data but without moving the iterator.