Change next_dissector mechanism
Add a new next_dissector() function on the flow dissector to rely on it
instead of implicit connection. This allow to give ownership of the flow
(thus the send) to the next dissector. As a result, the behavior of the
http dissector (an example of flow dissector) is the same as any other
dissector.