Fix Doxygen image issue with Breathe
Images in the Doxygen doc are not copied to the Doxygen output dir when
creating the xml files. Breathe needs this to be able to include it
correctly in the Sphinx output. As a fix, we force Doxygen to generate
both the xml and html output and put both results in the same folder.