Add grammar recursion
Add define("element") to allow predefinition of element.
It create a GrammarProxy that will then be resolved to "element" rule.

At compilation time we check if compound type are making recursive call. If so,
we create a DGRecurs element that is linked to compound by next() but that will
allow compilation to continue adding element after itself (on _next).

At runtime, DGRecurs will add its normal next (not recursion) to a stack _recurs
defined on ParseContext. Then if parse() reach an end point it will look for a
next element in _recurs stack and continue with it.