Fix invalid iterator when adding an header after a deleted one
Array is splitting buffer after each element in order to avoid loosing iterator
on next buffer if some of the element are removed. Nevertheless, if an array end
with a Token the iterator pointing to the end of the element is already at the
start of a new buffer (created by a split done by Token).

Consequently, when an element of such an array is deleted the element preceding
have its corrensponding subbuffer (made of 2 iteraror) broken. Its end iterator
was pointing to the next buffer.