Avoid use of lua state from an ffi call
We can get access to the lua state even if ffi doesn't provide it. We
were using it for state machine timeout. So we move timeout callback
into a run interrupt (that is called from a classic cfunc call in
mainloop).

This commit break support for C API on state machine. See
lib/haka/state_machine.c:399