Fix random crash
If the stream lua object changes (for instance with the stream field of
tcp_stream, if the gc runs it might destroy the lua object which will be
replaced by a new one if this same field is used again later), the
coroutine manager crashes as its inner stream (stored as a weak ref)
will be nil.

To avoid the problem, the stream is no longer stored inside the
coroutine manager.